Bug #1656: Lineage graph is clipped with latest ant 4.x version (#559)
* typefy SampleDetailsLineage and ref

* fix useTimeline

* make sample Partial

* fix viewport collapse

* add type support for dagre

* add type support for freezeman-d3-graph

* correct GraphConfiguration

* simplify ternary operation

* fix lineage with css

* linageStyle to lineageStyle

* more precise typing

* move overflow style into react

* use external style instead

* fix previous commit
