+++ frontend/package.json
@@ -38,7 +38,7 @@
     "debounce": "^1.2.0",
     "dompurify": "^2.2.7",
     "dotenv": "^10.0.0",
-    "freezeman-d3-graph": "^2.6.2",
+    "freezeman-d3-graph": "^2.6.5",
     "fuse.js": "^6.4.0",
     "jsondiffpatch": "^0.4.1",
     "jwt-decode": "^2.2.0",
@@ -67,6 +67,7 @@
     "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.18.6",
     "@reduxjs/toolkit": "^1.9.1",
+    "@types/dagre": "^0.7.48",
     "babel-loader": "^8.1.0",
     "babel-plugin-import": "^1.13.3",
     "babel-polyfill": "^6.26.0",
