+++ frontend/src/components/containers/ContainersDetailContent.js
@@ -20,10 +20,6 @@ const pageStyle = {
   overflow: "hidden",
 }
 
-const tabsStyle = {
-  marginTop: 8,
-}
-
 const tabStyle = {
   padding: "0 24px 24px 24px",
   overflow: "auto",
@@ -77,7 +73,7 @@ const ContainersDetailContent = ({
             </Space>
         } />
       <PageContent loading={!isLoaded && isFetching} style={pageStyle}>
-        <Tabs activeKey={activeKey} onChange={setActiveKey} size="large" type="card" style={tabsStyle}>
+        <Tabs activeKey={activeKey} onChange={setActiveKey} size="large" type="card">
           <TabPane tab="Overview" key="overview" style={tabStyle}>
             <Descriptions bordered={true} size="small">
               <Descriptions.Item label="ID" span={2}>{container.id}</Descriptions.Item>
