+++ frontend/src/components/samples/details/SampleDetailsContent.js
@@ -63,14 +63,16 @@ const pageStyle = {
   overflow: "hidden",
 }
 
-const tabsStyle = {
-  marginTop: 8,
-}
-
 const tabStyle = {
   padding: "0 24px 24px 24px",
   overflow: "auto",
   height: "100%",
+}
+
+const lineageStyle = {
+  ...tabStyle,
+  overflow: 'hidden',
+  height: '90vh'
 }
 
 const listSampleMetadata = (token, options) =>
@@ -198,7 +200,7 @@ const SampleDetailsContent = ({
       {error &&
         <ErrorMessage error={error} />
       }
-      <Tabs activeKey={activeKey} onChange={setActiveKey} size="large" type="card" style={tabsStyle}>
+      <Tabs activeKey={activeKey} onChange={setActiveKey} size="large" type="card" className={(activeKey === 'lineage' ? 'lineage-tab-active' : '')}>
         <TabPane tab="Overview" key="overview" style={tabStyle}>
           <Descriptions bordered={true} size="small">
             <Descriptions.Item label="ID">{sample.id}</Descriptions.Item>
@@ -352,7 +354,7 @@ const SampleDetailsContent = ({
           </Descriptions>
         </TabPane>
 
-        <TabPane tab={`Lineage`} key="lineage" style={tabStyle}>
+        <TabPane tab={`Lineage`} key="lineage" style={lineageStyle}>
           <SampleDetailsLineage sample={sample} 
             handleSampleClick={navigateToSample}
             handleProcessClick={navigateToProcess}
