+++ frontend/src/utils/ref.tsx
@@ -3,18 +3,18 @@ import { useRef, useCallback, useEffect, useState } from 'react';
 /**
  * https://gist.github.com/DominicTobias/c8579667e8a8bd7817c1b4d5b274eb4c
  * */
-export function useResizeObserver(width, height) {
+export function useResizeObserver(width: number, height: number) {
   const ResizeObserver = window.ResizeObserver;
   const [size, setSize] = useState({ width, height });
-  const resizeObserver = useRef(null);
+  const resizeObserver = useRef<ResizeObserver | null>(null);
 
-  const onResize = useCallback(entries => {
+  const onResize: ResizeObserverCallback = useCallback((entries) => {
     const { width, height } = entries[0].contentRect;
     setSize({ width, height });
   }, []);
 
-  const ref = useCallback(
-    node => {
+  const ref: React.LegacyRef<HTMLElement> = useCallback(
+    (node: HTMLElement) => {
       if (node !== null) {
         if (resizeObserver.current) {
           resizeObserver.current.disconnect();
@@ -23,11 +23,11 @@ export function useResizeObserver(width, height) {
         resizeObserver.current.observe(node);
       }
     },
-    [onResize]
+    [ResizeObserver, onResize]
   );
 
   useEffect(() => () => {
-    resizeObserver.current.disconnect();
+    resizeObserver.current?.disconnect();
   }, []);
 
   return { ref, size };
