+++ frontend/src/utils/useTimeline.ts
@@ -4,17 +4,17 @@ const OFFSET = 350
 
 export default function useTimeline() {
 
-  const [timelineWidth, setTimelineWidth] = useState(0);
+  const [timelineWidth, setTimelineWidth] = useState<number>(0);
   const timelineMarginLeft = -timelineWidth + OFFSET;
-  const timelineRef = useRef(null);
+  const timelineRef = useRef<HTMLElement | null>(null);
 
   useEffect(() => {
     if (timelineRef.current)
       setTimelineWidth(timelineRef.current.clientWidth || window.innerWidth - OFFSET)
-  });
+  }, []);
   useEffect(() => {
     const resizeListener = () => {
-      setTimelineWidth(timelineRef.current.clientWidth)
+      setTimelineWidth((timelineWidth) => timelineRef.current?.clientWidth ?? timelineWidth)
     };
     window.addEventListener('resize', resizeListener);
     return () => window.removeEventListener('resize', resizeListener);
